Move BalanceInfo to primitives
